feat: add business rules for author and publisher deletion

+++ src/main/java/com/exemplo/livraria/service/EditoraService.java
@@ -4,7 +4,9 @@ import com.exemplo.livraria.dto.EditoraRequestDTO;
 import com.exemplo.livraria.dto.EditoraResponseDTO;
 import com.exemplo.livraria.entity.Editora;
 import com.exemplo.livraria.exception.EditoraNaoEncontradaException;
+import com.exemplo.livraria.exception.NegocioException;
 import com.exemplo.livraria.repository.EditoraRepository;
+import com.exemplo.livraria.repository.LivroRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,12 +15,14 @@ import java.util.List;
 public class EditoraService {
 
     private EditoraRepository editoraRepository;
+    private LivroRepository livroRepository;
 
-    public EditoraService(EditoraRepository editoraRepository) {
+    public EditoraService(EditoraRepository editoraRepository, LivroRepository livroRepository) {
         this.editoraRepository = editoraRepository;
+        this.livroRepository = livroRepository;
     }
 
-    private EditoraResponseDTO editoraToDTO(Editora editora){
+    private EditoraResponseDTO editoraToDTO(Editora editora) {
         EditoraResponseDTO editoraDTO = new EditoraResponseDTO();
         editoraDTO.setNome(editora.getNome());
         editoraDTO.setId(editora.getId());
@@ -27,20 +31,20 @@ public class EditoraService {
     }
 
     // será usado no GET e retorna uma lista de ResponseDTO
-    public List<EditoraResponseDTO> listarTodos(){
+    public List<EditoraResponseDTO> listarTodos() {
         List<Editora> lista = editoraRepository.findAll();
         return lista.stream().map(this::editoraToDTO).toList();
     }
 
     // será usado no GET/id e retorna um objeto ResponseDTO
-    public EditoraResponseDTO listarPorID(Long id){
+    public EditoraResponseDTO listarPorID(Long id) {
         Editora editora = editoraRepository.findById(id)
-                .orElseThrow( () -> new EditoraNaoEncontradaException("Editora não encontrada"));
+                .orElseThrow(() -> new EditoraNaoEncontradaException("Editora não encontrada"));
         return editoraToDTO(editora);
     }
 
     // será usado no POST e retorna um ResponseDTO
-    public EditoraResponseDTO cadastrar(EditoraRequestDTO editoraRequestDTO){
+    public EditoraResponseDTO cadastrar(EditoraRequestDTO editoraRequestDTO) {
         Editora editora = new Editora();
         editora.setNome(editoraRequestDTO.getNome());
         editora.setCidade(editoraRequestDTO.getCidade());
@@ -48,17 +52,22 @@ public class EditoraService {
         return editoraToDTO(editora);
     }
 
-    // será usado no DELETE e não possui retorno
-    public void deletar(Long id){
+// será usado no DELETE e não possui retorno
+    public void deletar(Long id) {
+        if (livroRepository.existsByEditoraId(id)) {
+            throw new NegocioException("Não é possível excluir uma editora que possui livros vinculados a ela.");
+        }
         Editora editoraParaDeletar = editoraRepository.findById(id)
-                .orElseThrow(()-> new EditoraNaoEncontradaException("Editora não encontrada"));
+                .orElseThrow(() -> new EditoraNaoEncontradaException("Editora não encontrada"));
         editoraRepository.delete(editoraParaDeletar);
+
+
     }
 
     // será usado no PUT e retorna um ResponseDTO
-    public EditoraResponseDTO atualizar(EditoraRequestDTO editoraDTO, Long id){
+    public EditoraResponseDTO atualizar(EditoraRequestDTO editoraDTO, Long id) {
         Editora editoraParaAtualizar = editoraRepository.findById(id)
-                .orElseThrow( () -> new EditoraNaoEncontradaException("Editora não encontrada"));
+                .orElseThrow(() -> new EditoraNaoEncontradaException("Editora não encontrada"));
         editoraParaAtualizar.setCidade(editoraDTO.getCidade());
         editoraParaAtualizar.setNome(editoraDTO.getNome());
         editoraRepository.save(editoraParaAtualizar);
